Improve parsing of escape sequences in identifiers and URIs.
Implement proper parsing of escape sequences in unquoted URL function
arguments, and ensure escape sequences in quoted URLs are properly
translated when creating a PropertyValuePart.

The secondary parsing in PropertyValuePart to assign a type string
can fail when the identifier contains escape sequences.  Pass in a
`hint` object based on the original token emitted from the lexer to
disambiguate in these cases.