Validate `font-family` and the `font` compound property.
This uses the Matcher framework to handle a fairly-complex shorthand
property.

This patch is based on a patch set by Onno van der Zee, originally
submitted as PR #164.  Using the Matcher framework allowed the
original implementation to be greatly simplified, but many of the
test cases are from the original patch set.